server/requirements/prod.txt
Flask
Flask-RESTful
kafka-python
requests
flask-socketio